SendReceiveNonParallel
