+++ src/libraries/System.Net.Sockets/tests/FunctionalTests/SendReceive/SendReceiveListener.cs
@@ -1,17 +1,8 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System.Collections.Generic;
-using System.Linq;
-using System.Runtime.InteropServices;
-using System.Threading;
 using System.Threading.Tasks;
-using Microsoft.DotNet.RemoteExecutor;
-using Microsoft.DotNet.XUnitExtensions;
 using Xunit;
-using Xunit.Abstractions;
-using Xunit.Sdk;
-
 
 namespace System.Net.Sockets.Tests
 {
